+++ vvp/vvp_net.h
@@ -52,7 +52,6 @@ class  vvp_net_fun_t;
 class  vvp_net_fil_t;
 
 /* Core net function types. */
-class  vvp_fun_concat;
 class  vvp_fun_drive;
 class  vvp_fun_part;
 
@@ -1366,55 +1365,6 @@ class vvp_net_fil_t  : public vvp_vpi_callback {
 
 /* **** Some core net functions **** */
 
-/* vvp_fun_concat
- * This node function creates vectors (vvp_vector4_t) from the
- * concatenation of the inputs. The inputs (4) may be vector or
- * vector8 objects, but they are reduced to vector4 values and
- * strength information lost.
- *
- * The expected widths of the input vectors must be given up front so
- * that the positions in the output vector (and also the size of the
- * output vector) can be worked out. The input vectors must match the
- * expected width.
- */
-class vvp_fun_concat  : public vvp_net_fun_t {
-
-    public:
-      vvp_fun_concat(unsigned w0, unsigned w1,
-		     unsigned w2, unsigned w3);
-      ~vvp_fun_concat();
-
-      void recv_vec4(vvp_net_ptr_t port, const vvp_vector4_t&bit,
-                     vvp_context_t context) final;
-
-      void recv_vec4_pv(vvp_net_ptr_t port, const vvp_vector4_t&bit,
-			unsigned base, unsigned vwid, vvp_context_t) final;
-    private:
-      unsigned wid_[4];
-      vvp_vector4_t val_;
-};
-
-class vvp_fun_concat8  : public vvp_net_fun_t {
-
-    public:
-      vvp_fun_concat8(unsigned w0, unsigned w1,
-		     unsigned w2, unsigned w3);
-      ~vvp_fun_concat8();
-
-      void recv_vec4(vvp_net_ptr_t port, const vvp_vector4_t&bit,
-                     vvp_context_t context) final;
-      void recv_vec8(vvp_net_ptr_t port, const vvp_vector8_t&bit) final;
-
-      void recv_vec4_pv(vvp_net_ptr_t port, const vvp_vector4_t&bit,
-			unsigned base, unsigned vwid, vvp_context_t) final;
-      void recv_vec8_pv(vvp_net_ptr_t p, const vvp_vector8_t&bit,
-			unsigned base, unsigned vwid) final;
-
-    private:
-      unsigned wid_[4];
-      vvp_vector8_t val_;
-};
-
 /*
  * The vvp_fun_force class objects are net functors that use their input
  * to force the associated filter. They do not actually  have an
@@ -1434,29 +1384,6 @@ class vvp_fun_force : public vvp_net_fun_t {
       void recv_vec4(vvp_net_ptr_t port, const vvp_vector4_t&bit,
 		     vvp_context_t context);
       void recv_real(vvp_net_ptr_t port, double bit, vvp_context_t);
-};
-
-/* vvp_fun_repeat
- * This node function create vectors by repeating the input. The width
- * is the width of the output vector, and the repeat is the number of
- * times to repeat the input. The width of the input vector is
- * implicit from these values.
- */
-class vvp_fun_repeat  : public vvp_net_fun_t {
-
-    public:
-      vvp_fun_repeat(unsigned width, unsigned repeat);
-      ~vvp_fun_repeat();
-
-      void recv_vec4(vvp_net_ptr_t port, const vvp_vector4_t&bit,
-                     vvp_context_t context);
-      void recv_vec4_pv(vvp_net_ptr_t port, const vvp_vector4_t&bit,
-			unsigned int base, unsigned int vwid,
-			vvp_context_t context) final;
-
-    private:
-      unsigned wid_;
-      unsigned rep_;
 };
 
 /* vvp_fun_drive
